chore: refactor CN managers in dedicated module
