+++ core/control-plane/control-plane-core/build.gradle.kts
@@ -25,6 +25,7 @@ dependencies {
     implementation(project(":core:common:boot"))
     implementation(project(":core:control-plane:control-plane-catalog"))
     implementation(project(":core:control-plane:control-plane-contract"))
+    implementation(project(":core:control-plane:control-plane-contract-manager"))
     implementation(project(":core:control-plane:control-plane-transfer"))
     implementation(project(":core:control-plane:control-plane-aggregate-services"))
     implementation(project(":core:common:lib:util-lib"))
